feat: 사용자 예약 페이지 반환 API 구현
