+++ README.md
@@ -81,3 +81,12 @@
   ```http request
   HTTP/1.1 204
   ```
+
+# 3단계
+
+-[ ] /reservation 요청 시 사용자 예약 페이지를 응답한다
+-[ ] 인기 테마 조회 기능을 추가한다
+    -[ ] / 요청 시 인기 테마 페이지를 응답한다
+    -[ ] 최근 일주일을 기준으로 하여 해당 기간 내에 방문하는 예약이 많은 테마 10개를 조회한다
+- [ ] 사용자는 날짜와 테마를 선택하면 예약 가능한 시간을 확인할 수 있다
+    - [ ] 이미 예약된 시간의 경우 isAlreadyBooked를 true로 설정하여 응답한다
