+++ src/main/java/roomescape/controller/UserPageController.java
@@ -8,31 +8,16 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class PageController {
+public class UserPageController {
     @GetMapping("/")
     public ResponseEntity<Void> getWelcomePage() {
         HttpHeaders header = new HttpHeaders();
-        header.setLocation(URI.create("/admin"));
+        header.setLocation(URI.create("/reservation"));
         return new ResponseEntity<>(header, HttpStatus.PERMANENT_REDIRECT);
     }
 
-    @GetMapping("/admin")
-    public String getMainPage() {
-        return "admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
+    @GetMapping("/reservation")
     public String getReservationPage() {
-        return "admin/reservation-new";
-    }
-
-    @GetMapping("/admin/time")
-    public String getTimePage() {
-        return "admin/time";
-    }
-
-    @GetMapping("/admin/themes")
-    public String getThemePage() {
-        return "admin/theme";
+        return "reservation";
     }
 }
